requirements.txt
langchain_openai
langchain_core
langchain_community
streamlit
python_dotenv
langchain_ollama
fastapi
uvicorn
sse_starlette
langserve